==> notebooks/3.0-building-predictive-model.ipynb <==
# Building Predictive Models - Part 1

import pandas as pd
import numpy as np
import os

### Import Data

# set the path of processed data
processed_data_path = os.path.join(os.path.pardir, 'data', 'processed')
train_file_path = os.path.join(processed_data_path, 'train.csv')
test_file_path = os.path.join(processed_data_path, 'test.csv')

train_df = pd.read_csv(train_file_path, index_col = 'PassengerId')
test_df = pd.read_csv(test_file_path, index_col = 'PassengerId')

train_df.info()

test_df.info()

### Data Preparation

X = train_df.loc[:, 'Pclass':].values.astype('float')
y = train_df['Survived'].ravel()
print(X.shape, y.shape)

# train test split
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 0)
print(X_train.shape, y_train.shape)
print(X_test.shape, y_test.shape)

# average survival in train and test
print('mean survival in train: {0:0.3f}'.format(np.mean(y_train)))
print('mean survival in test: {0:0.3f}'.format(np.mean(y_test)))

#### Check Scikit-Learn Version

import sklearn
sklearn.__version__

Make sure you have Scikit-Learn v0.19 or newer. Otherwise update it and restart kernel.

#!conda update -y scikit-learn

### Baseline Model

# import function
from sklearn.dummy import DummyClassifier

# create baseline model
model_dummy = DummyClassifier(strategy = 'most_frequent', random_state = 0)

# train model
model_dummy.fit(X_train, y_train)

# default score - accuracy
print('score for baseline model: {0:0.2f}'.format(model_dummy.score(X_test, y_test)))

# performance metrics
from sklearn.metrics import accuracy_score, confusion_matrix, precision_score, recall_score

# accuracy score
print('accuracy for baseline model: {0:0.2f}'.format(accuracy_score(y_test, model_dummy.predict(X_test))))

# confusion matrix
print('confusion matrix for baseline model: \n {0}'.format(confusion_matrix(y_test, model_dummy.predict(X_test))))

# precision and recall scores
print('precision for baseline model: {0:0.2f}'.format(precision_score(y_test, model_dummy.predict(X_test))))
print('recall for baseline model: {0:0.2f}'.format(recall_score(y_test, model_dummy.predict(X_test))))

### First Kaggle Submission

# converting to matrix
test_X = test_df.values.astype('float')

# get predictions
predictions = model_dummy.predict(test_X)

df_submission = pd.DataFrame({ 'PassengerId': test_df.index, 'Survived': predictions })
df_submission.head()

submission_data_path = os.path.join(os.path.pardir, 'data', 'external')
submission_file_path = os.path.join(submission_data_path, '01_dummy.csv')
print(submission_file_path)

df_submission.to_csv(submission_file_path, index = False)

def get_submission_file(model, filename, message):
    # converting to matrix
    test_X = test_df.values.astype('float')
    # make predictions
    predictions = model.predict(test_X)
    # submission dataframe
    df_submission = pd.DataFrame({ 'PassengerId': test_df.index, 'Survived': predictions })
    # submission file
    submission_data_path = os.path.join(os.path.pardir, 'data', 'external')
    submission_file_path = os.path.join(submission_data_path, filename)
    # write to file
    df_submission.to_csv(submission_file_path, index = False)
    # submit using Kaggle API
    from subprocess import call
    call(['kaggle', 'competitions', 'submit', 'titanic', '-f', submission_file_path, '-m', message])

# get submission file
get_submission_file(model_dummy, '01_dummy.csv', 'Baseline Model Submission')

### Logistic Regression Model

# import function
from sklearn.linear_model import LogisticRegression

# create model
model_lr_1 = LogisticRegression(random_state = 0, solver = 'liblinear')

# train model
model_lr_1.fit(X, y)

# evaluate model
print('score for logistic regression - version 1: {0:0.2f}'.format(model_lr_1.score(X_test, y_test)))

# performance metrics
# accuracy
print('accuracy for logistic regression - version 1: {0:0.2f}'.format(accuracy_score(y_test, model_lr_1.predict(X_test))))
# confusion matrix
print('confusion matrix logistic regression - version 1: \n {0}'.format(confusion_matrix(y_test, model_lr_1.predict(X_test))))
# precision
print('precision for logistic regression - version 1: {0:0.2f}'.format(precision_score(y_test, model_lr_1.predict(X_test))))
# recall
print('recall for logistic regression - version 1: {0:0.2f}'.format(recall_score(y_test, model_lr_1.predict(X_test))))

# model coefficients
model_lr_1.coef_

### Second Kaggle Submission

# get submission file
get_submission_file(model_lr_1, '02_lr.csv', 'Logistic Regression Model v1 Submission')

# Building Predictive Models - Part 2

### Hyperparameter Optimization

# base model
model_lr = LogisticRegression(random_state = 0, solver = 'liblinear')

# import function
from sklearn.model_selection import GridSearchCV

# create parameter grid
parameters = {
    'C': [ 0.1, 0.2, 0.5, 1.0, 10.0, 50.0, 100.0, 1000.0 ],
    'penalty': [ 'l1', 'l2' ]
}
clf = GridSearchCV(model_lr, param_grid = parameters, cv = 5) # 5-fold crossvalidation

clf.fit(X, y)

clf.best_params_

print('best score: {0:0.2f}'.format(clf.best_score_))

# evaluate model
print('score for logistic regression - version 2: {0:0.2f}'.format(clf.score(X_test, y_test)))

### Third Kaggle Submission

# get submission file
get_submission_file(clf, '03_lr.csv', 'Logistic Regression Model v2 Submission')

### Feature Normalization and Standardization

# import function
from sklearn.preprocessing import MinMaxScaler, StandardScaler

#### Feature Normalization

# feature normalization
scaler = MinMaxScaler()
X_scaled = scaler.fit_transform(X)

X_scaled[:,0].min(), X_scaled[:,0].max()

# normalize test data
X_test_scaled = scaler.transform(X_test)

#### Create Model after Normalization

model_lr = LogisticRegression(random_state = 0, solver = 'liblinear')
parameters = {
    'C': [ 0.1, 0.2, 0.5, 1.0, 10.0, 50.0, 100.0, 1000.0 ],
    'penalty': [ 'l1', 'l2' ]
}
clf = GridSearchCV(model_lr, param_grid = parameters, cv = 5)
clf.fit(X_scaled, y)

clf.best_params_

clf.best_score_

# evaluate model
print('score for logistic regression - after normalization: {0:0.2f}'.format(clf.score(X_test_scaled, y_test)))

#### Submit to Kaggle

# get submission file
get_submission_file(clf, '04_lr.csv', 'Logistic Regression Model (Post Normalization) Submission')

#### Feature Standardization

# feature standardization
scaler = StandardScaler()
X_scaled = scaler.fit_transform(X)
X_test_scaled = scaler.transform(X_test)

#### Create Model after Standardization

model_lr = LogisticRegression(random_state = 0, solver = 'liblinear')
parameters = {
    'C': [ 0.1, 0.2, 0.5, 1.0, 10.0, 50.0, 100.0, 1000.0 ],
    'penalty': [ 'l1', 'l2' ]
}
clf = GridSearchCV(model_lr, param_grid = parameters, cv = 5)
clf.fit(X_scaled, y)

clf.best_params_

clf.best_score_

# evaluate model
print('score for logistic regression - after standardization: {0:0.2f}'.format(clf.score(X_test_scaled, y_test)))

#### Submit to Kaggle

# get submission file
get_submission_file(clf, '05_lr.csv', 'Logistic Regression Model (Post Standardization) Submission')

### Model Persistance

# import pickle library
import pickle

# create file paths
model_file_path = os.path.join(os.path.pardir, 'models', 'lr_model.pkl')
print(model_file_path)
scaler_file_path = os.path.join(os.path.pardir, 'models', 'lr_scaler.pkl')
print(scaler_file_path)

# open files to write
model_file_pickle = open(model_file_path, 'wb')
scaler_file_pickle = open(scaler_file_path, 'wb')

# persist the model and scaler
pickle.dump(clf, model_file_pickle)
pickle.dump(scaler, scaler_file_pickle)

# close the files
model_file_pickle.close()
scaler_file_pickle.close()

#### Load Persisted Files

# open files in read mode
model_file_pickle = open(model_file_path, 'rb')
scaler_file_pickle = open(scaler_file_path, 'rb')

# load model and scaler
clf_loaded = pickle.load(model_file_pickle)
scaler_loaded = pickle.load(scaler_file_pickle)

# close the files
model_file_pickle.close()
scaler_file_pickle.close()

clf_loaded

scaler_loaded

X_test_scaled = scaler_loaded.transform(X_test)
print('score for persisted logistic regression: {0:0.2f}'.format(clf_loaded.score(X_test_scaled, y_test)))
