==> notebooks/4.0-building-machine-learning-api.ipynb <==
# Hello World API with Flask

import os
hello_world_script_file = os.path.join(os.path.pardir, 'src', 'models', 'hello_world_api.py')

%%writefile $hello_world_script_file
from flask import Flask, request

app = Flask(__name__)

@app.route('/api/hello', methods = ['POST'])
def say_hello():
    data = request.get_json(force = True)
    name = data['name']
    return 'Hello {0}!'.format(name)

if __name__ == '__main__':
    app.run(port = 10001, debug = True)

import json
import requests

# invoke hello world API
url = 'http://127.0.0.1:10001/api/hello'
data = json.dumps({ 'name': 'Jack' })
r = requests.post(url, data)

print(r.text)

# Machine Learning API with Flask

### Building API

machine_learning_api_script_file = os.path.join(os.path.pardir, 'src', 'models', 'machine_learning_api.py')

%%writefile $machine_learning_api_script_file
from flask import Flask, request
import pandas as pd
import json
import pickle
import os

app = Flask(__name__)

# Load model and scaler files
model_path = os.path.abspath('models')
model_file_path = os.path.join(model_path, 'lr_model.pkl')
scaler_file_path = os.path.join(model_path, 'lr_scaler.pkl')

model_file = open(model_file_path, 'rb')
scaler_file = open(scaler_file_path, 'rb')
model = pickle.load(model_file)
scaler = pickle.load(scaler_file)
model_file.close()
scaler_file.close()

columns = [
    u'Age', u'Fare', u'FamilySize', u'IsMother', u'IsMale', u'IsAdult',
    u'Deck_A', u'Deck_B', u'Deck_C', u'Deck_D', u'Deck_E', u'Deck_F', u'Deck_G', u'Deck_Z',
    u'Pclass_1', u'Pclass_2', u'Pclass_3',
    u'Title_Lady', u'Title_Master', u'Title_Miss', u'Title_Mr', u'Title_Mrs', u'Title_Officer', u'Title_Sir',
    u'Fare_Bin_very_low', u'Fare_Bin_low', u'Fare_Bin_high', u'Fare_Bin_very_high',
    u'Embarked_C', u'Embarked_Q', u'Embarked_S'
]

@app.route('/api', methods = ['POST'])
def make_prediction():
    # read json object and convert to json string
    data = json.dumps(request.get_json(force = True))
    # create pandas dataframe using the json string
    df = pd.read_json(data)
    # extract PassengerIds
    passenger_ids = df['PassengerId'].ravel()
    # actual survived values
    actuals = df['Survived'].ravel()
    # extract required columns and convert to matrix
    X = df[columns].values.astype('float')
    # transform the input
    X_scaled = scaler.transform(X)
    # make predictions
    predictions = model.predict(X_scaled)
    # create response dataframe
    df_response = pd.DataFrame({
        'PassengerId': passenger_ids,
        'Predicted': predictions,
        'Actual': actuals
    })
    return df_response.to_json()

if __name__ == '__main__':
    # host flask app at port 10001
    app.run(port = 10001, debug = True)

### Invoking API using Requests

import os
import pandas as pd
import numpy as np

processed_data_path = os.path.join(os.path.pardir, 'data', 'processed')
train_file_path = os.path.join(processed_data_path, 'train.csv')
train_df = pd.read_csv(train_file_path)

survived_passengers = train_df[train_df.Survived == 1][:5]
survived_passengers

def make_api_request(data):
    import requests
    # url for api
    url = 'http://127.0.0.1:10001/api'
    # make post request
    r = requests.post(url, data)
    return r.json()

make_api_request(survived_passengers.to_json())

# compare results
result = make_api_request(train_df.to_json())
df_result = pd.read_json(json.dumps(result))
df_result.head()

# accuracy level
np.mean(df_result.Predicted == df_result.Actual)
